@@ -45,7 +45,7 @@
     "@types/jest": "^29.5.2",
     "@types/lodash": "^4.17.0",
     "@types/multer": "^1.4.11",
-    "@types/node": "^20.3.1",
+    "@types/node": "^20.12.7",
     "@types/supertest": "^6.0.0",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
@@ -59,9 +59,9 @@
     "supertest": "^6.3.3",
     "ts-jest": "^29.1.0",
     "ts-loader": "^9.4.3",
-    "ts-node": "^10.9.1",
+    "ts-node": "^10.9.2",
     "tsconfig-paths": "^4.2.0",
-    "typescript": "^5.1.3"
+    "typescript": "^5.4.5"
   },
   "jest": {
     "moduleFileExtensions": [
