@@ -7,7 +7,11 @@ const routes: Routes = [
   {
     path: '',
     component: CompanysPage
-  }
+  },
+  // {
+  //   path: 'modal-visit-info',
+  //   loadChildren: () => import('./modal-visit-info/modal-visit-info.module').then( m => m.ModalVisitInfoPageModule)
+  // }
 ];
 
 @NgModule({
